requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
google-cloud-aiplatform==1.51.0
google-generativeai==0.4.1 